+++ administrator/components/com_gamification/views/goals/tmpl/default.php
@@ -10,7 +10,7 @@
 // no direct access
 defined('_JEXEC') or die;
 ?>
-<form action="<?php echo JRoute::_('index.php?option=com_gamification&view=achievements'); ?>" method="post" name="adminForm" id="adminForm">
+<form action="<?php echo JRoute::_('index.php?option=com_gamification&view=goals'); ?>" method="post" name="adminForm" id="adminForm">
     <?php if(!empty( $this->sidebar)): ?>
     <div id="j-sidebar-container" class="span2">
         <?php echo $this->sidebar; ?>
@@ -23,7 +23,7 @@ defined('_JEXEC') or die;
     // Search tools bar
     echo JLayoutHelper::render('joomla.searchtools.default', array('view' => $this));
     ?>
-        <table class="table table-striped" id="achievementsList">
+        <table class="table table-striped" id="goalsList">
             <thead><?php echo $this->loadTemplate('head'); ?></thead>
             <tfoot><?php echo $this->loadTemplate('foot'); ?></tfoot>
             <tbody><?php echo $this->loadTemplate('body'); ?></tbody>